\name{bulkQC-package}
\alias{bulkQC-package}
\docType{package}
\title{Quality Control and Outlier Identification in Bulk for Multicenter Trials}
\description{
Multicenter randomized trials involve the collection and analysis of data from numerous study participants across multiple sites. Outliers may be present. To identify outliers, this package examines data at the individual level (univariate and multivariate) and site-level (with and without covariate adjustment). Methods are outlined in further detail in Rigdon et al (to appear). 
}

\details{
\tabular{ll}{
Package: \tab bulkQC \cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2024-04-24\cr
License: \tab GPL-3\cr
}
}

\author{Joseph Rigdon \email{jrigdon@wakehealth.edu}}

\references{

Tukey J. Exploratory Data Analysis. 1st edition. Reading, Mass: Pearson; 1977. 712 p. 

Cortes D. Explainable outlier detection through decision tree conditioning. arXiv:200100636 [cs, stat] [Internet]. 2020 Jan 2 [cited 2021 Nov 12]; Available from: http://arxiv.org/abs/2001.00636 

Yang D, Dalton JE. A unified approach to measuring the effect size between two groups using SAS. 2012;6 

}
